added definition of 'GTOP_GUILE_LIBS' and 'GTOP_GUILE_INCS':
* gtopConf.sh.in: added definition of 'GTOP_GUILE_LIBS'
and 'GTOP_GUILE_INCS':

'GTOP_LIBS' and 'GTOP_INCS' now contains everything
that is needed to link with libgtop.

'GTOP_GUILE_LIBS' and 'GTOP_GUILE_INCS' now contains
everything that is needed to link with libgtop and its
guile interface. When guile cannot be found on the system
or building of the guile interface was disables, they are
identical to 'GTOP_LIBS' and 'GTOP_INCS'
[FIXME: Should I make them empty in this case?]