Fix timeout issue in DBus interface
Currently there was an infinite loop used to keep the packet
capture ongoing, thereby not allowing the control to return
out of the function well within the time .

Instead `while(1)` is replaced by invoking the loop's content
in fixed intervals of time by ` g_timeout_add ` and the capture
function is separated into two functions, viz.
  -  init_setup : setup of processes and pcap handles are fetched
  -  glibtop_init_packet_capture : having the content of the loop
                                   as earlier
Henceforth to access the packet capture API from libgtop there is
a provision to either keep running an endless packet capture which
shall be used in case of future proposed daemon approach
but if this is to be used as a part of another program then you
have to call the capture glibtop_init_packet_capture at fixed
intervals of time.