Check if sbrk is available
The only use of sbrk in libgtop is in an example which always calls it
with 0. Since the use of sbrk provides no functionality, just skip these
calls when sbrk is not available.

This problem was first reported on FreeBSD Bugzilla because FreeBSD
decides not to support sbrk on ARM64 and RISC-V:
https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=221751

Fix https://gitlab.gnome.org/GNOME/libgtop/issues/46