Use G_*_FORMAT when formatting 64-bit types
Several of the messages libgtop formats include guint64 values. 7e9fed15
("Use correct format specifiers") changed some of these from %llu to
%lu, but neither is correct on all platforms -- for example, compiling
the current version on 32-bit ARM produces lots of warnings.

The right way to do this is to use the G_*_FORMAT macros, which are
glib's equivalents of the standard inttypes.h PRI* macros.

There are also instances of a pid_t being printed. Cast these to gint64
(pid_t has to be a signed type, and there are already systems with
64-bit PIDs).