Ensure null termination on buffer
Whilst we were taking care to avoid writing past the end of the buffer,
we then assumed that the result would be a null terminated string. This
made it possible for us to end up reading past the end of the buffer
when using certain functions.

Avoid this by initialising the buffer to zeros, and only reading len-1
bytes, thus ensuring that whatever is read is followed by at least one
null byte.